working on new datasets

+++ judge/dataset_eval/judge-instruct.py
@@ -181,7 +181,7 @@ def main(args):
                 pending_data[idx]["rating"] = float(rating)
                 pending_data[idx]["judgement_pending"] = False
                 pending_data[idx]["rated_by"] = judge_model
-                if int(rating) < 6:
+                if float(rating) < 6:
                     print("------------------------------------------------------------------------------")
                     print(text)
                     print("****")
